+++ src/main/java/com/drmangotea/tfmg/content/machinery/misc/winding_machine/WindingMachineBlock.java
@@ -2,7 +2,10 @@ package com.drmangotea.tfmg.content.machinery.misc.winding_machine;
 
 import com.drmangotea.tfmg.base.TFMGShapes;
 import com.drmangotea.tfmg.registry.TFMGBlockEntities;
+import com.drmangotea.tfmg.registry.TFMGItems;
+import com.simibubi.create.content.kinetics.deployer.DeployerFakePlayer;
 import com.simibubi.create.content.kinetics.base.HorizontalKineticBlock;
+import com.simibubi.create.content.kinetics.deployer.DeployerFakePlayer;
 import com.simibubi.create.foundation.block.IBE;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -55,7 +58,7 @@ public class WindingMachineBlock extends HorizontalKineticBlock implements IBE<W
     @Override
     public void onRemove(BlockState pState, Level pLevel, BlockPos pPos, BlockState pNewState, boolean pIsMoving) {
         super.onRemove(pState, pLevel, pPos, pNewState, pIsMoving);
-        IBE.onRemove(pState,pLevel,pPos,pNewState);
+        IBE.onRemove(pState, pLevel, pPos, pNewState);
     }
 
     @Override
@@ -65,22 +68,25 @@ public class WindingMachineBlock extends HorizontalKineticBlock implements IBE<W
 
     @Override
     protected ItemInteractionResult useItemOn(ItemStack stack, BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hitResult) {
-
-
-        if(level.getBlockEntity(pos) instanceof WindingMachineBlockEntity be){
-            if(player.getItemInHand(hand).isEmpty()){
-                if(!be.inventory.isEmpty()&&!(be.inventory.getItem(0).getItem() instanceof SpoolItem)&&player.isShiftKeyDown()) {
+        if (level.getBlockEntity(pos) instanceof WindingMachineBlockEntity be) {
+            if (player.getItemInHand(hand).isEmpty()) {
+                if (!be.outputInventory.isEmpty()) {
+                    player.setItemInHand(hand, be.outputInventory.getStackInSlot(0));
+                    be.outputInventory.setItem(0, ItemStack.EMPTY);
+                    return ItemInteractionResult.SUCCESS;
+                }
+                if (!be.inventory.isEmpty() && !(be.inventory.getItem(0).getItem() instanceof SpoolItem) && !(player instanceof DeployerFakePlayer)) {
                     player.setItemInHand(hand, be.inventory.getStackInSlot(0));
                     be.inventory.setItem(0, ItemStack.EMPTY);
                     return ItemInteractionResult.SUCCESS;
                 }
-                if(!be.spool.isEmpty()){
-                    player.setItemInHand(hand, be.spool);
-                    be.spool = ItemStack.EMPTY;
+                if (player.isShiftKeyDown() && !be.getSpool().isEmpty() && be.getSpool().is(TFMGItems.EMPTY_SPOOL.get())) {
+                    player.setItemInHand(hand, be.getSpool());
+                    be.setSpool(ItemStack.EMPTY);
                     return ItemInteractionResult.SUCCESS;
                 }
-            }else {
-                if(be.inventory.isEmpty()&&!(player.getItemInHand(hand).getItem() instanceof SpoolItem)){
+            } else {
+                if (be.inventory.isEmpty() && !(player.getItemInHand(hand).getItem() instanceof SpoolItem)){
                     ItemStack stack1 = player.getItemInHand(hand).copy();
                     stack1.setCount(1);
                     be.inventory.setItem(0, stack1);